py-doc-backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
python-docx==1.1.2
typing_extensions==4.14.0
Werkzeug==3.1.3
